+++ src/vaping/plugins/vodka.py
@@ -1,7 +1,8 @@
+import copy
+
 import vaping
 import vaping.config
 import vaping.io
-import copy
 from vaping.plugins import PluginConfigSchema
 import confu.schema