+++ tests/test_fping.py
@@ -31,7 +31,11 @@ expect_verbose = {
         "loss": 0.333,
         "last": 50,
     },
-    "10.130.133.2 : - - -": {"host": "10.130.133.2", "cnt": 3, "loss": 1,},
+    "10.130.133.2 : - - -": {
+        "host": "10.130.133.2",
+        "cnt": 3,
+        "loss": 1,
+    },
     "example.com: Temporary failure in name resolution": {},
 }