+++ tests/test_fping_mtr.py
@@ -23,7 +23,11 @@ expect_verbose = {
         "cnt": 3,
         "loss": 0.333,
     },
-    "10.130.133.2 : - - -": {"host": "10.130.133.2", "cnt": 3, "loss": 1,},
+    "10.130.133.2 : - - -": {
+        "host": "10.130.133.2",
+        "cnt": 3,
+        "loss": 1,
+    },
     "example.com: Temporary failure in name resolution": {},
 }